fix: rolled back ProjectStatusSection
