+++ src/components/Project/ProjectBasics/ProjectForm/ProjectStatusSection.tsx
@@ -5,13 +5,12 @@ import { Control, UseFormGetValues } from 'react-hook-form';
 import { IProjectForm } from '@/interfaces/formInterfaces';
 import { useTranslation } from 'react-i18next';
 import { IOption } from '@/interfaces/common';
-import { isBefore, updateYear } from '@/utils/dates';
+import { getToday, isBefore, updateYear } from '@/utils/dates';
 import RadioCheckboxField from '@/components/shared/RadioCheckboxField';
 import ErrorSummary from './ErrorSummary';
-import { validateMaxNumber } from '@/utils/validation';
+import { getFieldsIfEmpty, validateMaxNumber } from '@/utils/validation';
 import _ from 'lodash';
 import { mapIconKey } from '@/utils/common';
-import { usePhaseValidation } from '@/hooks/usePhaseValidation';
 import { useAppSelector } from '@/hooks/common';
 import { selectProject } from '@/reducers/projectSlice';
 
@@ -37,12 +36,108 @@ const ProjectStatusSection: FC<IProjectStatusSectionProps> = ({
   const currentPhase = getValues('phase').value;
   const { t } = useTranslation();
 
-  const { validatePhase, phaseRequirements } = usePhaseValidation({
-    getValues,
-  });
-  const [proposalPhase, designPhase, constructionPhase] = useMemo(
-    () => phases.map(({ value }) => value),
-    [phases],
+  const [phaseRequirements, setPhaseRequirements] = useState<Array<string>>([]);
+
+  const [
+    proposalPhase,
+    designPhase,
+    programmedPhase,
+    draftInitiationPhase,
+    draftApprovalPhase,
+    constructionPlanPhase,
+    constructionWaitPhase,
+    constructionPhase,
+    warrantyPeriodPhase,
+    completedPhase,
+  ] = useMemo(() => phases.map(({ value }) => value), [phases]);
+
+  const validatePhase = useMemo(
+    () => ({
+      required: t('validation.required', { field: t('validation.phase') }) ?? '',
+      validate: {
+        isPhaseValid: (phase: IOption) => {
+          if (phase.value === '') {
+            return t('validation.required', { field: t('validation.phase') }) ?? '';
+          }
+
+          const phaseToSubmit = phase.value;
+          const programmed = getValues('programmed');
+          const fields: Array<string> = [];
+          const fieldsIfEmpty = (fields: Array<string>) => getFieldsIfEmpty(fields, getValues);
+
+          const programmedRequirements = [
+            'planningStartYear',
+            'constructionEndYear',
+            'category',
+            'masterClass',
+            'class',
+          ];
+          const planningRequirements = ['estPlanningEnd', 'estPlanningStart', 'personPlanning'];
+          const generalConstructionRequirements = [
+            'estConstructionStart',
+            'estConstructionEnd',
+            'personConstruction',
+          ];
+          const combinedRequirements = [
+            ...programmedRequirements,
+            ...planningRequirements,
+            ...generalConstructionRequirements,
+          ];
+
+          // Check fields that cannot be empty
+          switch (phaseToSubmit) {
+            case programmedPhase:
+              fields.push(...fieldsIfEmpty([...programmedRequirements]));
+              break;
+            case draftInitiationPhase:
+            case draftApprovalPhase:
+            case constructionPlanPhase:
+            case constructionWaitPhase:
+              fields.push(...fieldsIfEmpty([...programmedRequirements, ...planningRequirements]));
+              break;
+            case constructionPhase:
+              fields.push(...fieldsIfEmpty([...combinedRequirements, 'constructionPhaseDetail']));
+              break;
+            case warrantyPeriodPhase:
+            case completedPhase:
+              if (isBefore(getToday(), getValues('estConstructionEnd'))) {
+                return t('validation.phaseTooEarly', { value: phase.label });
+              }
+
+              fields.push(...fieldsIfEmpty([...combinedRequirements]));
+              break;
+          }
+
+          // Check if programmed has the correct value
+          if (phase.value === proposalPhase || phase.value === designPhase) {
+            if (programmed) {
+              fields.push('programmed');
+            }
+          } else {
+            if (!programmed) {
+              fields.push('programmed');
+            }
+          }
+          setPhaseRequirements(fields);
+
+          return fields.length === 0;
+        },
+      },
+    }),
+    [
+      t,
+      getValues,
+      proposalPhase,
+      designPhase,
+      programmedPhase,
+      draftInitiationPhase,
+      draftApprovalPhase,
+      constructionPlanPhase,
+      constructionWaitPhase,
+      constructionPhase,
+      warrantyPeriodPhase,
+      completedPhase,
+    ],
   );
 
   const validateConstructionPhaseDetails = useMemo(
@@ -58,12 +153,12 @@ const ProjectStatusSection: FC<IProjectStatusSectionProps> = ({
         },
       },
     }),
-    [constructionPhase, getValues, t],
+    [completedPhase, constructionPhase, getValues, t, warrantyPeriodPhase],
   );
 
   const isConstructionPhaseDetailsDisabled = useMemo(() => {
     return currentPhase !== constructionPhase;
-  }, [currentPhase, constructionPhase]);
+  }, [currentPhase, constructionPhase, warrantyPeriodPhase, completedPhase]);
 
   const validateProgrammed = useMemo(
     () => ({
